Cars in lanes + bigger cars
Cars now spawn inside the lanes, however they spawn on top of each other which is been marked on github issue #4, also changed size of cars and trucks, and made trucks y and x size adjustable via a variable. This was done so when you adjust the size of the truck size it also adjusts where it checks for collisions.